refactor: 로그인 관련 로직 함수 분리
[#397]
