+++ frontend/src/components/ui/compound/NavigationBar/Menu.tsx
@@ -6,7 +6,7 @@ import {
 } from '@heroicons/react/24/outline';
 import { css } from 'styled-components';
 
-import { serverStore } from '@stores/config/serverStore';
+import { getLoginUri } from '@utils/login';
 
 import Button from '@common/Button';
 import FlexBox from '@common/FlexBox';
@@ -16,10 +16,6 @@ import ServerStationFilters from '@ui/ServerStationFilters';
 import StationListWindow from '@ui/StationList/StationListWindow';
 import StationSearchWindow from '@ui/StationSearchWindow';
 import LogoIcon from '@ui/Svg/LogoIcon';
-
-import { SERVERS } from '@constants';
-
-import type { LoginUriResponse } from '@type/login';
 
 import { useNavigationBar } from './hooks/useNavigationBar';
 
@@ -52,14 +48,10 @@ const Menu = () => {
       </Button>
       <Button
         aria-label="로그인 하기"
-        onClick={async () => {
-          const mode = serverStore.getState();
-
-          const loginUriResponse = await fetch(
-            `${SERVERS[mode]}/oauth/google/login-uri?redirect-uri=http://localhost:3000/google`
-          ).then<LoginUriResponse>((response) => response.json());
-
-          window.location.href = loginUriResponse.loginUri.replace(/;/, '');
+        onClick={() => {
+          getLoginUri('google').then((loginUri) => {
+            window.location.href = loginUri;
+          });
         }}
       >
         <UserCircleIcon width="2.8rem" stroke="#333" />
